changed vcFileError to vcError, and moved error structs out of State into vcError
