--- src/vcFileError.h
@@ -1,8 +1,0 @@
-#ifndef vcFileError_h__
-#define vcFileError_h__
-
-#include "vcState.h"
-
-void vcFileError_AddError(vcState *pProgramState, const vcState::ErrorItem &error);
-
-#endif //vcFileError_h__
